refactor(deepgram): extract a transport-agnostic Deepgram Flux TTS base
DeepgramFluxTTSBase holds the Flux TTS protocol — query string, Speak,
Flush, Interrupt, Configure, message handling and settings — and reaches
its transport through _transport_send_json, _transport_is_active,
_connect and _disconnect. DeepgramFluxTTSService is now the websocket
implementation of that seam.

The STT base moves from flux/base.py to flux/stt_base.py, matching the
new flux/tts_base.py.
